+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.6 (2023-04-18)
+
+- Add Poland, Ukraine, Moldova, Greece, Turkey, India, China
+- Update unit tests
+- Add common service method for phone check
+- Update README.md
+
 ## 0.1.5 (2023-04-17)
 
 - Add Italy